done buy cart and now . checkout user

+++ uteshop/src/main/java/cuoiki/ltweb/dao/IShippingCompanyDAO.java
@@ -2,6 +2,7 @@ package cuoiki.ltweb.dao;
 
 public interface IShippingCompanyDAO {
 
-	public int getIdShippingCompanyByName(String companyname);
+
+	public long getIdShippingCompanyByName(String companyname);
 
 }
